Transitioned navbar into component, styling fixes
