+++ styles/globals.css
@@ -92,6 +92,7 @@ body {
       rgb(var(--background-end-rgb))
     )
     rgb(var(--background-start-rgb));
+  overflow-x: hidden;
 }
 
 a {
@@ -100,6 +101,17 @@ a {
 }
 
 h2 {
+}
+
+/* Hide scrollbar for Chrome, Safari and Opera */
+body::-webkit-scrollbar {
+  display: none;
+}
+
+/* Hide scrollbar for IE, Edge and Firefox */
+body {
+  -ms-overflow-style: none; /* IE and Edge */
+  scrollbar-width: none; /* Firefox */
 }
 
 /* GAMES */
@@ -166,11 +178,15 @@ h2 {
 }
 
 #nav-container {
-  width: 55%;
+  width: 100%;
   align-self: flex-end;
   display: flex;
   justify-content: space-between;
-  margin-right: 5%;
+  padding: 8px;
+  align-items: center;
+}
+
+#nav-links-container {
   background: linear-gradient(
     to right,
     #ffffff00,
@@ -180,14 +196,19 @@ h2 {
     #8787870c,
     #90909001
   );
-  padding: 8px;
+  width: 50%;
+  display: flex;
+  justify-content: space-between;
   align-items: center;
 }
 
 @media screen and (max-device-width: 850px) {
   #nav-container {
-    width: 85%;
     font-size: 0.8em;
+  }
+
+  #nav-links-container {
+    width: 80%;
   }
 
   .Home_center__4BFgC::before,
@@ -213,11 +234,23 @@ h2 {
 @media screen and (max-device-width: 800px) {
   #nav-container {
     font-size: 0.5em;
-    width: 100%;
   }
 
   #nav-container a {
     padding: 0;
+  }
+  #nav-links-container {
+    width: 85%;
+  }
+
+  #nav-logo-container {
+    width: 12%;
+    height: 5vh;
+  }
+
+  #ygg-logo {
+    width: auto;
+    height: 100%;
   }
 
   .card {
